refactor: LineTest에서 MockBooleanGenerator 클래스 분리
