+++ src/test/java/domain/generator/MockBooleanGenerator.java
@@ -1,0 +1,17 @@
+package domain.generator;
+
+import java.util.List;
+
+public class MockBooleanGenerator implements BooleanGenerator {
+    private final List<Boolean> values;
+    private int index = 0;
+
+    public MockBooleanGenerator(List<Boolean> values) {
+        this.values = values;
+    }
+
+    @Override
+    public boolean generate() {
+        return values.get(index++);
+    }
+}
